# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API tareas",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/api-tasks/public"
        }
    ],
    "paths": {
        "/api/task": {
            "get": {
                "summary": "Mostrar las tareas",
                "operationId": "6b555a1de0e16a13a2c81f8ba7ed89b8",
                "responses": {
                    "200": {
                        "description": "Muestra todas las tareas"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Put"
                ],
                "summary": "Edit Task",
                "description": "Edita una tarea",
                "operationId": "edit",
                "requestBody": {
                    "description": "Objeto tarea",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "nuevaTarea"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Se guardo Correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "{status:1,taks:{name:tarea,description:description}}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "New Task",
                "description": "Crea una nueva tarea",
                "operationId": "store",
                "requestBody": {
                    "description": "Objeto tarea",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "nuevaTarea"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Se creo una nueva",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "{status:1,taks:{name:tarea,description:description}}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Delete"
                ],
                "summary": "Delete Task",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "exito"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/task/{cant}": {
            "get": {
                "summary": "Mostrar las tareas dependiendo de la cantidad que se le mande",
                "operationId": "7c66357d83d2b1fc4e8c76e90d38fb68",
                "parameters": [
                    {
                        "name": "cant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un json del paginador."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "summary": "Mostrar usuarios",
                "operationId": "eaedf6aa84fde38a08bf5d87308169f7",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    }
                }
            }
        },
        "/api/users/client": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "new User",
                "description": "Crea un nuevo usuario",
                "operationId": "store",
                "requestBody": {
                    "description": "Objeto User",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Se guardo Correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "{user:{username:admin,password:admin,api_token:asdsa123}}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/users/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign in",
                "description": "Login by Username, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No Content"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}